Refactor Drawer into smaller components
